// src/apis/models/Volo_Abp_AuditLogging_EntityChangeWithUsernameDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Volo_Abp_AuditLogging_EntityChangeDto } from './Volo_Abp_AuditLogging_EntityChangeDto';
export type Volo_Abp_AuditLogging_EntityChangeWithUsernameDto = {
    entityChange?: Volo_Abp_AuditLogging_EntityChangeDto;
    userName?: string | null;
};